Rename classes and folder
