+++ twilio-authenticator-sample/src/main/java/com/twilio/authenticatorsample/appslist/MainActivity.java
@@ -1,6 +1,5 @@
-package com.twilio.authenticatorsample.main;
+package com.twilio.authenticatorsample.appslist;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
