+++ twilio-authenticator-sample/src/main/res/layout/app_bar_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.twilio.authenticatorsample.main.MainActivity">
+    tools:context="com.twilio.authenticatorsample.appslist.MainActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
