+++ twilio-authenticator-sample/src/main/res/layout/fragment_app_detail.xml
@@ -9,7 +9,7 @@
     android:gravity="center"
     android:orientation="vertical"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.twilio.authenticatorsample.main.AppDetailFragment">
+    tools:context="com.twilio.authenticatorsample.appdetail.AppDetailFragment">
 
     <TextView
         android:id="@+id/totp"
